Minor fixes  - Updated TODO  - cleaned up some more files  - removed some redundant references that were causing exceptions  - added support for setting the app name through an environment variable  - renamed wsgi.py to run.py to avoid confusion  - imported the asgi app in run.py to top level (api.lds:asgi --> api:asgi)
