+++ .flaskenv
@@ -1,2 +1,2 @@
-FLASK_APP=api\lds\wsgi.py
+FLASK_APP=api\lds\run.py
 FLASK_DEBUG=1
