+++ TODO.md
@@ -4,10 +4,15 @@
  - user password reset
  - user password change
  - app token reset
- - remove all redundant code/templates
+ - remove all redundant code/templates (only app/auth/routes.py remains)
  - permission and discord endpoints
  - steam account link functionality + endpoints
  - twitch account link functionality + endpoints
+
+
+**For next deployment**
+ - set env variable 'SRLM_APP_NAME'
+ - change supervisor file for hypercorn - app moved from api.lds:asgi to api:asgi
 
 
 **Future**
