+++ api/__init__.py
@@ -1,1 +1,3 @@
 import api.lds
+
+from api.lds import asgi
