+++ api/lds/app/auth/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 bp = Blueprint('auth', __name__)
 
-from api.lds.app.auth import routes, forms, email, functions
+from api.lds.app.auth import email, functions
